Fix calculation in finance_calculator.py

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-monthly_income = float (input("Enter your monthly income: "))
-monthly_expenditure = float (input("Enter your total monthly expenses: "))
+monthly_income = int (input("Enter your monthly income: "))
+monthly_expenditure = int (input("Enter your total monthly expenses: "))
 
 monthly_savings = monthly_income - monthly_expenditure
